major refactor
