+++ benches/compiler/Project.toml
@@ -1,0 +1,2 @@
+[deps]
+Cthulhu = "f68482b8-f384-11e8-15f7-abe071a5a75f"
